+++ swift/Sources/GildedRose/Extensions/UInt8+quality.swift
@@ -1,5 +1,5 @@
 //
-//  Int+quality.swift
+//  UInt8+quality.swift
 //
 //
 //  Created by Mauro Bianchelli on 2024-06-09.
@@ -8,7 +8,7 @@
 import Foundation
 
 /// This is not the best solution, but since the `Item` class cant be modified (for example to use a range instead of an integer) it is a workaround.
-public extension Int {
+public extension UInt8 {
     var qualityBoundsChecker: Self {
         if self >= 50 {
             return 50
